+++ test/validate.spec.php
@@ -1,0 +1,7 @@
+<?php
+
+describe('Krak Validation', function() {
+    describe('Validators', function() {
+        require_once __DIR__ . '/validators.php';
+    });
+});